add text format and colors utility refeactor.
