+++ src/index.js
@@ -5,8 +5,7 @@ import path from 'node:path';
 import { argv } from 'node:process';
 import minimist from 'minimist';
 import templateCategories from './templateList';
-import colors from './utils/colors';
-import setColor from './utils/setColor';
+import { colors, setColor } from './utils/colors';
 import initialize from './initialize';
 import { formatAppName, validateAppName } from './utils/appName';
 import helpMessage from './utils/helpMessage';
