+++ src/templateList.js
@@ -1,5 +1,4 @@
-import colors from './utils/colors';
-import setColor from './utils/setColor';
+import { colors, setColor, separator, setUnderlineOff} from './utils/colors';
 
 const templateCategories = {
   'Frontend': {
@@ -13,35 +12,29 @@ const templateCategories = {
     'React JavaScript': [
       {
         value: 'react',
-        title: setColor('React'),
-        description: colors.error('React + Vite.'),
+        title: `${setColor('React')}${separator('React')}${setUnderlineOff(colors.error('React + Vite.'))}`,
       },
       {
         value: 'react-dev',
-        title:  setColor('React Dev'),
-        description: colors.error('Tailwind + Linter + Tests'),
+        title:  `${setColor('React Dev')}${separator('React Dev')}${setUnderlineOff(colors.error('Tailwind + Linter + Tests'))}`,
       },
       {
         value: 'react-webpack',
-        title: setColor('React Webpack'),
-        description: colors.error('React + Webpack.')
+        title: `${setColor('React Webpack')}${separator('React Webpack')}${setUnderlineOff(colors.error('React + Webpack.'))}`,
       },
       {
         value: 'react-dev-webpack',
-        title: setColor('React Dev Webpack'),
-        description: colors.error('Tailwind + Linter + Tests')
+        title: `${setColor('React Dev Webpack')}${separator('React Dev Webpack')}${setUnderlineOff(colors.error('Tailwind + Linter + Tests'))}`,
       },
     ],
     'React TypeScript': [
       {
         value: 'react-typescript',
-        title:  setColor('React TypeScript'),
-        description: colors.error('React + Vite + TypeScript'),
+        title:  `${setColor('React TypeScript')}${separator('React TypeScript')}${setUnderlineOff(colors.error('React + Vite + TypeScript'))}`,
       },
       {
         value: 'react-dev-typescript',
-        title:  setColor('React Dev TypeScript'),
-        description: colors.error('Tailwind + Linter + Tests'),
+        title:  `${setColor('React Dev TypeScript')}${separator('React Dev TypeScript')}${setUnderlineOff(colors.error('Tailwind + Linter + Tests'))}`,
       },
     ]
   },
@@ -49,9 +42,8 @@ const templateCategories = {
     'NodeJS': [
       {
         value: 'node-express',
-        title:  setColor('Node Express'),
-        description: colors.error('Node + Express + Jest + Eslint.'),
-      }
+        title:  `${setColor('Node Express')}${separator('Node Express')}${setUnderlineOff(colors.error('Express + Linter + Tests'))}`,
+      },
     ]
   }
 }
